add duke package
