+++ src/main/java/duke/DeadlineCommand.java
@@ -1,3 +1,5 @@
+package duke;
+
 import java.io.IOException;
 
 public class DeadlineCommand extends Command {
