+++ src/main/java/duke/EventCommand.java
@@ -1,3 +1,5 @@
+package duke;
+
 import java.io.IOException;
 
 public class EventCommand extends Command {
