+++ src/main/java/duke/TodoCommand.java
@@ -1,3 +1,5 @@
+package duke;
+
 import java.io.IOException;
 
 public class TodoCommand extends Command {
