+++ src/main/java/duke/UnmarkCommand.java
@@ -1,19 +1,22 @@
+package duke;
+
 import java.io.IOException;
 import java.util.List;
 
-public class DeleteCommand extends Command {
+public class UnmarkCommand extends Command {
     private int taskNumber;
 
-    public DeleteCommand(int taskNumber) {
+    public UnmarkCommand(int taskNumber) {
         this.taskNumber = taskNumber;
     }
 
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws IOException {
         List<Task> taskList = tasks.getTasks();
-        Task taskToDelete = taskList.get(taskNumber - 1);
-        taskList.remove(taskToDelete);
+        Task currTask = taskList.get(taskNumber - 1);
+
+        currTask.markAsNotDone();;
         storage.save(taskList);
-        ui.showDeleteTask(taskToDelete, taskList);
+        ui.showTaskMarkUndone(currTask);
     }
 }
